Revert "32 put item endpoint"
