+++ backend/db/repository.go
@@ -91,7 +91,6 @@ type ItemRepository interface {
 	UpdateItemStatus(ctx context.Context, id int32, status domain.ItemStatus) error
 	GetItemStatus(ctx context.Context, id int32) (domain.ItemStatus, error)
 	SearchItemByName(ctx context.Context, keyword string) ([]domain.Item, error)
-	PutItem(ctx context.Context, item domain.Item,itemID int64) (domain.Item, error)
 }
 
 type ItemDBRepository struct {
@@ -109,18 +108,6 @@ func (r *ItemDBRepository) AddItem(ctx context.Context, item domain.Item) (domai
 	// TODO: if other insert query is executed at the same time, it might return wrong id
 	// http.StatusConflict(409) 既に同じIDがあった場合
 	row := r.QueryRowContext(ctx, "SELECT * FROM items WHERE rowid = LAST_INSERT_ROWID()")
-
-	var res domain.Item
-	return res, row.Scan(&res.ID, &res.Name, &res.Price, &res.Description, &res.CategoryID, &res.UserID, &res.Image, &res.Status, &res.CreatedAt, &res.UpdatedAt)
-}
-
-func (r *ItemDBRepository) PutItem(ctx context.Context, item domain.Item, itemID int64) (domain.Item, error) {
-	if _, err := r.ExecContext(ctx, "UPDATE items SET name = ?, price = ?, description = ?, category_id = ?, seller_id = ?, status = ? WHERE ID = ?", item.Name, item.Price, item.Description, item.CategoryID, item.UserID, item.Status,itemID); err != nil {
-		return domain.Item{}, err
-	}
-	// TODO: if other insert query is executed at the same time, it might return wrong id
-	// http.StatusConflict(409) 既に同じIDがあった場合
-	row := r.QueryRowContext(ctx, "SELECT * FROM items WHERE ID = ?",itemID)
 
 	var res domain.Item
 	return res, row.Scan(&res.ID, &res.Name, &res.Price, &res.Description, &res.CategoryID, &res.UserID, &res.Image, &res.Status, &res.CreatedAt, &res.UpdatedAt)
