--- frontend/simple-mercari-web/src/components/EditItem/index.tsx
@@ -1,1 +1,0 @@
-export * from './EditItem';
