--- frontend/simple-mercari-web/src/components/UserItemList/index.tsx
@@ -1,1 +1,0 @@
-export * from "./UserItemList";
