+++ frontend/simple-mercari-web/src/components/UserProfile/UserProfile.tsx
@@ -4,7 +4,6 @@ import { useParams } from "react-router-dom";
 import { MerComponent } from "../MerComponent";
 import { toast } from "react-toastify";
 import { ItemList } from "../ItemList";
-import { UserItemList } from "../UserItemList";
 import { fetcher } from "../../helper";
 
 interface Item {
@@ -104,7 +103,7 @@ export const UserProfile: React.FC = () => {
 
           <div>
             <h2>Item List</h2>
-            {<UserItemList items={items} />}
+            {<ItemList items={items} />}
           </div>
         </div>
       </div>
